% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowering_data.R
\docType{data}
\name{flowering_data}
\alias{flowering_data}
\title{Made-up floral observation data}
\format{
A data frame with 36 rows and 7 variables:
\describe{
\item{species}{A made-up species name.}
\item{year}{Year of observation (used for day-of-year calculations, accounting for leap years).}
\item{flower_start}{First day of year when an individual plant was observed in flower.}
\item{flower_end}{Last day of year when an individual plant was observed in flower.}
\item{median_flowers}{Median number of flowers per plant on that day estimate}
\item{lower_sd}{If flowers are estimated, lower}
\item{upper_sd}{If flowers are estimated, upper}
}
}
\usage{
data(flowering_data)
}
\description{
A dataset containing 40 made-up first and last flower dates for fictional species.
}
\keyword{datasets}
