% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowering_data.R
\docType{data}
\name{flowering_data}
\alias{flowering_data}
\title{made up floral observation data}
\format{
a dataframe with 36 rows and 7 variables.
}
\usage{
data(flowering_data)
}
\description{
A dataset with 40 made up first and last flower dates for made up species
}
\details{
\itemize{
\item 'species' is a made up name.
\item 'year' of observations, required to cater day of year (DOY) to leap years.
\item 'flower_start'. first day of year which a phenological event, in this case flowering, was observed on an individual plant.
\item 'flower_end' final day of year which the individual was observed to be in flower.
}
}
\keyword{datasets}
